+++ .idea/stylesheetLinters/stylelint.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="StylelintConfiguration">
+    <config-file value="$PROJECT_DIR$/doctor_appointment_frontend/.stylelintrc.json" />
+  </component>
+</project>
