+++ doctor_appointment_frontend/package.json
@@ -46,6 +46,8 @@
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "@babel/preset-react": "^7.22.5",
     "eslint-config-airbnb": "^19.0.4",
+    "stylelint": "^15.10.3",
+    "stylelint-config-standard": "^34.0.0",
     "tailwindcss": "^3.3.3"
   }
 }
