--- doctor_appointment_frontend/src/components/doctors.js
@@ -1,34 +1,0 @@
-const list = {
-    doctors: [
-        {
-            id: 1,
-            name: 'Dr. John Doe',
-            speciality: 'Dentist',
-            image: '../assets/images/doc1.jpg',
-            description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quos.',
-        },
-        {
-            id: 2,
-            name: 'Dr. Alec Patel',
-            speciality: 'Pediatrician',
-            image: '../assets/images/doc2.jpg',
-            description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quos.',
-        },
-        {
-            id: 3,
-            name: 'Dr. Jean mac',
-            speciality: 'Neurologist',
-            image: '../assets/images/doc3.jpg',
-            description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quos.',
-        },
-        {
-            id: 4,
-            name: 'Dr. Mary Jane',
-            speciality: 'Cardiologist',
-            image: '../assets/images/doc4.jpg',
-            description: 'Lorem ipsum dolor sit amet consectetur adipisicing elit. Quisquam, quos.',
-        },
-    ],
-};
-
-export default list;
