+++ doctor_appointment_frontend/src/css/home.css
@@ -34,11 +34,6 @@
   overflow: hidden;
 }
 
-
-
 .reservation {
   background-color: rgba(150, 191, 1, 0.9);
-
 }
-
-
