+++ doctor_appointment_frontend/tailwind.config.js
@@ -7,5 +7,4 @@ module.exports = {
     extend: {},
   },
   plugins: [],
-}
-
+};
